6-5控制风扇完成

+++ 6-4PWM驱动舵机/Hardware/Servo.h
@@ -2,6 +2,6 @@
 #define __SERVO_H
 
 void Servo_Init(void);
-void Servo_SetAngle(float Angle);
+void Servo_SetAngle(uint16_t Angle);
 
 #endif
